requirements.txt
Django>=4.2,<5.0
djangorestframework
boto3
requests
python-dotenv
gunicorn
django-cors-headers
